pyproject.toml
[project]
name = "seo-pipeline"
version = "0.1.0"
description = "SEO monitoring pipeline for capetowndata.com"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    "httpx[http2]>=0.27",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "python-dotenv>=1.0",
    "google-api-python-client>=2.120",
    "google-auth>=2.28",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "respx>=0.21",
]
django = [
    "django>=5.0",
]

[project.scripts]
seo-pipeline = "seo_pipeline.run:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/seo_pipeline"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"
